Create README.md
Myself Benedict Maria, a dedicated CSE student, is on a journey of self-improvement, focusing on enhancing my soft skills. Outside of academics, I'm find solace in the pages of books and the melodies of music. I'm fluent in English , known for self-motivation and strong work ethic and driving towards success in both my studies and personal endeavors.